Replace CI `RUST_MIN_VER_PKGS` with `--workspace --ignore-private`. (#782)
The CI `RUST_MIN_VER_PKGS` configuration variable was meant to track
packages that are intended for publishing in order to limit MSRV checks
to only those packages. This same goal can be achieved universally
without the need for a configuration variable via the `--ignore-private`
`cargo-hack` option.

The immediate effect on this repo is that `xilem_core` is no longer part
of the MSRV check because it has `publish = false`. That is fine,
because the MSRV check will run just before publishing, triggered by the
PR that will remove `publish = false`, ensuring that the MSRV is
accurate before publishing.