Improve IME area reporting and set IME area in `TextArea` (#785)
This adds the required context methods for (re)setting the IME area.
These are added to the mutable context types, which I believe is the
right place.

`TextArea` now uses `PlainEditor::ime_cursor_area` to report its IME
area.

Note Masonry currently reports new IME cursor areas only when the widget
itself moves or the focus changes to a new widget, but Parley's
`PlainEditor` determines the IME cursor area based on the selection and
preedit state of the editor, meaning the area can change based on events
as well. Other widgets may choose to have different behavior still.

(It appears Masonry currently doesn't account for the IME area
potentially changing based on relayout, which may be a bug.)

This replaces the old IME area reporting with a check at the end of the
rewrite pass to see if IME is active, and if it is, whether the IME area
has moved. If it has, the new area is reported to the platform.

---------

Co-authored-by: Daniel McNab <36049421+DJMcNab@users.noreply.github.com>