masonry: Remove `focus` from `PointerState`
This is unused. It was added to Druid in
<https://github.com/linebender/druid/pull/842>.

In the `winit` world, this would be done instead by customizing
the window creation with `WindowAttributesExtMacOS` and the
`with_accepts_first_mouse` function. (So it would no longer be
something that you have to check on every mouse click event as
it is part of the `WinitView` within `winit`.