Virtual Scrolling Widget (#882)
This is a minimum viable virtual scrolling widget, making the following
assumptions:
-  It does not *care* about scrollbars. This has several big advantages:
    - It allows for "infinite" virtual up and down-scrolling
- It allows for much more control by users of how the scrolling happens
(e.g. a date based scrollbar)

It does *not* assume that each item has a fixed height, and instead uses
some best-effort heuristics to work out which items should be loaded.

We've done some thinking about how to handle arbitrary transforms of
child widgets. Essentially, the only way we can reasonably handle this
is for each child widget to be automatically clipped to their layout
rectangle (...maybe e.g. 20% larger than it; that's hard because if we
don't use fixed margins then the variable scroll controller needs to
know about it).

This also does not handle focus and other details around (e.g.)
textboxes properly; it's not actually possible for the VirtualScroll
widget to know which child has the focus, so it can't keep that widget
around. Saving the textbox state would be nice, but requires a redesign
of a lot of things, essentially.

![A screenshot of fizzbuzz starting at 'FizzBuzz' at the top of the
screen, followed by 42031, going to
42106.](https://github.com/user-attachments/assets/fe55be05-ad07-4dc7-b7c2-c9209f5a114b)

Fixes #51

---------

Co-authored-by: Bruce Mitchener <bruce.mitchener@gmail.com>