Update to `keyboard-types` 0.8 (#940)
With W3C specs, I think we should be mapping meta / super differently
than we were before. This will change again with winit 0.31 where the
names of those keys on the winit side will have changed. (And also, much
of this particular mapping won't be necessary.)

The bulk of this change is that in `keyboard-types` 0.8, the `Key`
enumeration has changed to have a separate enumeration for `NamedKey`
values.

---------

Co-authored-by: Mads Marquart <mads@marquart.dk>