Prepare to release 0.3.0. (#973)
We skip Xilem 0.2.0 and go straight to 0.3.0 in order to synchronize the
version with its companion packages that are expected to be updated in
lock-step for now.

[The one exception is
`tree_arena`](https://xi.zulipchat.com/#narrow/channel/419691-linebender/topic/Xilem.20version.20sync/with/517174026),
which will retain its own version pacing, because it's likely to see
fewer updates.