xilem: support disabling inputs (#977)
This commit firstly introduces a breaking change by changing the
WidgetPod constructors from taking an Affine to taking a WidgetOptions
containing an Affine so that then the disabled boolean can be added to
the WidgetOptions struct, allowing widgets to be created in disabled
state. The Xilem views for the 3 currently existing input widgets then
also gain a `disabled` method in order to change the disabled state.