Implement reparenting in tree_arena (#984)
This commit firstly implements root_ids and child_ids iterators so that
the reparenting implementation can be properly tested.

Reparenting is then implemented in both the safe and the unsafe tree.
For the unsafe tree no new unsafe code is introduced.

To make the tests more readable a new Node struct is introduced in the
test code so that the tree after the reparenting can be compared with a
single assert_eq!.