masonry: fix panic on event when widget is not in the tree (#999)
If a widget is removed from the tree in response to an event it can
cause a panic in the `on_event` function. I have seen this in a Xilem
app when a `text_box(..).on_enter(..)` handler causes the text_box and
its parent to be removed from the view. I have also seen it when a
button responds to an accessibility click event causing it to be
removed. I'm not certain if it is a subsequent event (like "key up")
being delivered to the removed widget that tiggers the panic or if it is
traversing the ancestor list for the same event when the ancestor has
been removed.