Take a WidgetPod for the root widget (#1061)
This does enable properties to be set for the root widget but the
motivation for making this change is that it paves the way for removing
the RootWidget widget.
(WindowView uses an AnyWidgetView for the main widget currently and
AnyWidgetView::build returns a Pod, so without this change we couldn't
forward the generated widget id).