Remove `Adapt` for expediency (#1085)
The [adapt](https://docs.rs/xilem_core/latest/xilem_core/fn.adapt.html)
`View` has use cases around message handling, especially around
immutable data structures. See also [Raph's blog
post](https://raphlinus.github.io/rust/gui/2022/05/07/ui-architecture.html#finer-grained-change-propagation-memoizing).
However, the API currently committed is incompatible with #1079.

This PR adds `map_message`, a view which handles the use-case of
handling every possible message (which as far as I can tell, was only an
accidental benefit of adapt, but what most users were using it for).
It's clear that our message model still needs work, but that isn't a
priority at the moment.

Outside of this, the API of `adapt` post #1079 isn't obvious to me. None
of our current examples are using its interesting power (which is
accessing the state you passed to the child after the child runs).
Because of this, I don't think redesigning it is worth the time it will
take to do so for me at the moment, but having it in its panicky state
isn't great.

My intention is to make an issue to redesign and restore this (which I
don't plan to work on), if this is approved.

---------

Co-authored-by: Philipp Mildenberger <philipp@mildenberger.me>