Placehero: Perform a styling pass on the status UI (#1092)
![image](https://github.com/user-attachments/assets/9291ce95-82c1-4864-b49c-bb53363986a6)

Note that this doesn't include #1091 and #1089, which are still in
review.

Note also the artifact at the top right of the screenshot. Those are
pre-existing, which I haven't taken the time to track down yet. I
believe it's related to the scrollbar, but I'm not sure how. I'm also
not sure where the grey colour is coming from.