xilem_core: Add a `WithoutElements` `ViewSequence` (#608)
This is similar as the `fork` view, but it allows using
`View`s/`ViewSequence`s without elements within a `ViewSequence` context
that contains elements, see
https://github.com/linebender/xilem/issues/461#issuecomment-2282857369
for the original motivation.

Taking `xilem_web` as example, this should be possible (as well as
something similar as in the linked comment):

```rust
fn run_once_fragment() -> impl DomFragment<()> {
    without_elements(run_once(|| {}))
}
```

The `WidgetViewSequence` is "fixed" as well, by using the "right" widget
type (`DynWidget` is only used in `AnyView`, whereas I think `Box<dyn
Widget>` is more general and correct).

---------

Co-authored-by: Daniel McNab <36049421+DJMcNab@users.noreply.github.com>