Add a flex row helper to Xilem (#1102)
This is a little papercut that I often run into, that the flex direction
being defined after the sequence means that it can be hard to understand
what element a piece of code relates to.

This is especially prevalent in nested flexes (e.g. a vertical flex
containing horizontal flexes), as the changing flow direction is
important to understand.