Placehero: Add support for showing a single thread (#1114)
On the timeline (there is now a `View Replies` button on each post):

![image](https://github.com/user-attachments/assets/5fc42a93-5004-4ae5-89f4-405ec89da730)

In a thread (i.e. once the button has been pressed):

![image](https://github.com/user-attachments/assets/86d57972-91cd-408f-8312-91ab5d03df2b)

Note that to get this image, I followed two "View Replies"; the first to
get to [this post](https://mastodon.online/@raph/114530193780235287).

Note a few features:
- There is now a `⬅️ Back to Timeline` button in the sidebar. This jumps
you back to Raph's timeline.
- The ancestors and descendants of the "thread" are now visible.
- There is a "replies" label after the "main" post in the new thread

A few caveats (which I'm treating as out of scope for this PR):
- The "continuity" of replies is not visible, i.e. which thread items
are actually replies to the previous item or one before.
- The scroll position is not handled very well; every page change jumps
to the top.
- See also [#masonry > Scrolling on (tab)
Focus](https://xi.zulipchat.com/#narrow/channel/317477-masonry/topic/Scrolling.20on.20.28tab.29.20Focus/with/526109923)
- The "replies" label is not great design