Cache tracing spans (#1132)
This is implementing the idea discussed in
https://github.com/linebender/xilem/pull/687#discussion_r1808472248

The reason I didn't do this before was that I thought you needed to get
the parents right. However, having re-read tracing's docs since then,
I've discovered that this isn't necessary.

As discussed in https://github.com/bevyengine/bevy/pull/9390, this can
get you pretty close to removing all the overhead; it's possible that
#687 could be undone.