Commit to stash-based implementation
When/if stash is replaced with a better method, this can be updated to
match.

Also, remove a TODO on how to select a new active child when the active
one is removed: if the caller wants a specific behavior, they can easily
implement it themselves by calling `set_active_child`.