Create WidgetArenaMut type and start refactor of pass internals.
This is the first step of a refactor where widgets, widget states and widget properties will be stored in the same object.
Doing so without making an absolutely unreadable diff is tough, so I split the work into several PRs.
This one just moves code around to make future diffs shorter.