Add active status (#1189)
This replaces "has pointer capture" with "active" as the name of the
concept of "this button / checkbox / widget is being interacted with".
Pointer capture is its own thing and in the future we might want them to
mean slightly different things.

(Also, when we add multiple pointers, checking an "active" flag will be
easier than checking "is this widget the pointer capture target of any
of the currently live pointers".)