DRAFT: Add WidgetTag
Add a new WidgetTag type, which lets you access widgets without having to store their WidgetIds and lets you skip downcasting.
Tweak the to-do-list example to demonstrate how to use those tags.

Tags are similar to Selectors back in Druid and have a similar purpose: to carry type information between places in the code.

This is a first step towards adding a more generalized query system to Masonry.