Create `NewWidget` type and refactor all methods that add a child widget to use that type (#1215)
This is a fairly large PR, with a few motivations:

- To introduce the concept of a `NewWidget` type which holds all the
options a widget will be created with.
- To standardize how new widgets are added, and remove the code
duplication in `with_child`, `with_child_id`, `with_child_pod`,
`insert_child`, `insert_child_pod`, etc, methods.
- To make the Xilem-side code much simpler by taking care of the widget
creation options on the Masonry side.
- To prepare the way for #1212.
- To prepare the upcoming lifecycle refactor, in which NewWidget will be
the only way to add a widget to the widget tree.

Overall I think it makes the Masonry APIs more elegant, at the cost of
making the builder-style "Create a widget tree with the children already
filled in" widget construction more verbose.
I think that's a worthwhile trade-off because (1) we still expect most
of our users to be Xilem-style frameworks which can tolerate the
verbosity (2) I expect the builder-style constructors to be
significantly altered in the lifecycle refactor (though I think this PR
is still worth adding even if the lifecycle refactor doesn't come).