Add rudimentary clipboard support (#1237)
This only supports copy-pasting strings, not rich text. And the way
Ctrl+V is detected is somewhat dubious.

But it does effectively add clipboard support to Masonry apps.

Ideally, on the long term, we may want to be able to programmatically
request clipboard contents; I'm not sure how to implement that cleanly
given RenderRoot's architecture, without making masonry_core depend
directly on a clipboard-handling library.

---------

Co-authored-by: Daniel McNab <36049421+DJMcNab@users.noreply.github.com>