Xilem: Make virtual scroll implementation much saner (#1260)
This is possible due to changes such as
https://github.com/linebender/xilem/pull/1079 and indeed partially
motivated that change.

One possibility here is to remove the `state_changed` boolean entirely.
However, I think that would actually still be useful in things like
`lens` (or indeed, in `virtual_scroll` itself, to avoid rerunning
callbacks in response to only a request rebuild). But this is an area of
semantics which probably needs more thought.

I have tested this in `virtual_cats`, and don't see any regressions.