DRAFT: Add layer system to Masonry
This adds a very rudimentary layer system to Masonry.

This is still a draft.
A more complete version would add a few RenderRootSignal variants for adding new layers.

Also, adding a layer stack kind of muddies the concept of a "root widget", and the APIs and doc don't really account for that. I'd like to fix that before a merge.