Rename some variables in Flex code to make it more readable.
I've often been tripped up by the "major/minor" names in the Flex code.
As far as I can tell, it's not a well-known naming scheme (e.g. Taffy doesn't use it, other GUI projects don't use it), and it's not intuitive to a newcomer.

This uses more explicit names, and documents their meaning.

Also renames self.direction to self.axis, mostly for conciseness.

Follow-up to #1276.