Remove focus chain and refactor tab-focus (#1226)
The new behavior has a few advantages:
- Avoids building and storing O(N) arrays of WidgetIds per-widget.
- When the user clicks somewhere and then presses `Tab`, the focused
widget is more likely to be close to whatever the user clicked.
- Overall more reliable than the old focus code.

Incidentally fixes a bug where a textbox sometimes woulnd't be shown as
focused.

Add unit tests for new behavior.
Add documentation for new behavior and new concept.

This PR introduces the concept of a "focus anchor", which is currently
under-specified and slightly inconsistent in corner cases. Future PRs
should give it more precise semantics.