When a new font is loaded, relayout everything (#1321)
This avoids issues where the initial render (say) doesn't have a given
font, but the font is later loaded.

This is always the case with when Xilem accesses the font context, but I
think it's also just generally the correct behaviour here.

This also fixes a minor issue with `request_render_all`, where it didn't
take into account `post_paint`.