Make `assert_any`, `assert_all` and `assert_none` not be macros (#1338)
Follow-up to https://github.com/linebender/xilem/pull/1335

The macros added in that PR were not using the fact that they are
macros. This would be confusing to users, as macros are to be avoided
unless necessary.

If in future we want to add assertion message support, etc, we can
always make them macros again (reasonably trivially)