Masonry: Give driver control over the Checkbox state (#1366)
This makes things easier for Xilem, as it controls the source of truth
for being checked. This operates somewhat similarly to `VirtualScroll`,
and also addresses the issues found in
https://github.com/linebender/xilem/pull/265.
I think it better reflects Masonry's low-level nature.

Working on this has uncovered another issue, which is that the tests
aren't properly updated since #1314.
That PR added a large indicator, even when hovered.
However, those indicators are inflated to draw outside of the widget's
area, so aren't captured here.
I don't have the motivation to fix this at the moment, but @tannal if
you are willing to that would be great. A similar "flex" approach as in
the new test from #1314 would be fine; just applying that to both tests
(this should likelyalso use `WidgetTag`).

I'd also like to see a test of checkbox with the Padding property,
because I'm not sure the rendering is correct, since `paint` doesn't
take it into account. (Either for the focused indicator, or for the
"box" position.
I suppose we should open an issue for that?