Optimize and fix window resizing. (#1371)
The Masonry `Window` struct was being created with a `0x0` size and a
`1.0` scale factor. Initial layout and render passes ran with that. Then
there would be an event for scale change (but still zero size) and
another pass. Then there would be a size change event and useful work
would finally take place.

This PR removes all that early useless work and creates the `Window`
struct immediately with the correct info.

---

A bunch of places would ask Winit for the window size, including the
`render` function. At least on Windows this would cause Winit to execute
a syscall. That is needless overhead, especially in a hot path, because
we already store that info in the `RenderRoot` struct.

This PR exposes that size info.

---

Masonry has been suffering from crashes on Windows 11 whenever a window
would be minimized or otherwise resized to zero (#979).

This PR resolves that issue. Fixes #979.

First, I refactored the redraw code so that there is a single source of
truth for the behavior. There was already divergence with the first
frame not doing the `AnimFrame` event and more issues could easily
follow with such a duplicated approach.

The redraw refactoring also introduced on-demand surface creation. This
is useful to actually solve the Windows crash issue, as now we can just
delete the old surface when a zero sized resize event comes in. Whenever
a reasonable size redraw happens, the surface will be recreated.

---

I have tested this as working nicely on Windows 11. Trying out crazy
resizings (especially to zero!), minimizing, and suspending on other
platforms would be very welcome.