Improve the setup of `masonry_core`'s default tracing backend (#1427)
This PR aims to improve the automatic setup of `masonry_core`'s default
tracing subscriber.

The Problem it tries to fix is that before the default tracing
subscriber would be setup regardless of if a user had already set a
tracing subscriber. This would introduce side effects like creating a
temp-file in "/tmp/" and logging a message to the user about it even
though the temp-file would never be used.

The new version introduces a call to a
`verify_subscriber_has_not_been_set()` check function before the default
tracing subscriber setup is attempted. It uses an undocumented tracing
function `tracing_core::dispatcher::has_been_set()` internally. The only
caveat with this approach is that this function returns true if any
tracing subscriber has been set before on the current thread regardless
of if it was the global default or if it is even active at the moment.

The assumption is that users who had set a tracing subscriber manually
will not rely on a default being set by the library. If they still wish
to use `masonry_core`'s default subscriber they get access to it through
the newly exposed `default_tracing_subscriber` function.