Allow non-contiguous app state in Xilem (#1444)
See the proposal in [#xilem > A mirage of a future
Xilem](https://xi.zulipchat.com/#narrow/channel/354396-xilem/topic/A.20mirage.20of.20a.20future.20Xilem/with/548204242).

This is a reasonably small addition to Xilem's model, but with the
potential to radically improve Xilem's usability. In it, we allow
multiple references

The biggest downside identified so far is the required new `()`
parameter to `lens`. This is due to
https://doc.rust-lang.org/error_codes/E0502.html. This is obviously
pretty bad, but I don't think it's a showstopper.
At the very least, I'd argue that Xilem needs an experiment like this to
happen; our development velocity isn't high enough that post-facto
reverting this would be unviable.

Status, of what needs to be ported:

- [X] Xilem Core
- [X] Xilem Core Examples/Tests
- [x] Xilem
- [x] Xilem Examples
- [x] Xilem Web

I'm proposing that these are left as follow-ups:

- [ ] Tests which use the new capabilities this provides
- [ ] New views for local variables
- [ ] Porting whichever examples this would be suitable for to use this
- [ ] The extension of the environment system into shared references.

Guidance for porting:

- Porting components is pretty trivial. Basically, any `View<State>` (or
`WidgetView<State>`) needs to be changed to `View<Edit<State>>` (or
`WidgetView<Edit<State>>`).
- Porting View implementations is also reasonably simple. We need:
  - new bounds that the `State` implements `ViewArgument`.
- Parameters to `View` functions of the form `&mut State` need to be
changed to `Arg<'_, State>`.
- Uses of `app_state` need to replaced with `State::reborrow_mut(&mut
app_state)`; in some cases, this can be skipped.