Remove double boxing of `ErasedAction` in `<Context>::submit_untyped_action` (#1452)
Remove the wrapping of an `ErasedAction` in a `Box::new(...)` in
`<Context>::submit_untyped_action` because `ErasedAction` is already the
`Box<dyn ...>` value that we are looking for in
`RenderRootSignal::Action(...)`.

I guess this was overlooked before because the
`<Context>::submit_untyped_action` method is used nowhere in the
project.