Add root padding parameter to TestHarness (#1449)
This solves a problem where a lot of screenshots are kind of hard to
parse because they're really tight (especially annoying in doc
screenshots).

Add some padding to one of the button screenshots.

I expect to add padding to a lot of screenshots, but I'd rather do it in
a follow-up PR to avoid committing to many new images to the repository.