Refactor xilem_core module structure
The overall goal of this PR is to ensure that major types/traits are declared in modules of the same name, and that module names reflect their most important item.
This also avoids "kitchen sink" modules which are hard to nagivate.

- Split View impls into `views::impl_xxx` modules.
- Split ViewSequence impls into `view_sequences::impl_xxx` modules.
- Make any_view a submodule of `views` .
- Split off `WithoutElements` into its own module.
- Split off `ElementSplice` into its own module.
- Split off `ViewPathTracker` into its own module.
- Rename `context` to `message_context`.
- Rename `deferred` to `message_proxy`.
- Rename `state` to `view_argument`.
- Rename `sequence` to `view_sequence`.

This is a move-only PR with no logic changes.