Add screenshot padding in `assert_render_snapshot` instead of layout (#1457)
Alternative to #1449.

This feels much nicer to me than the mechanism currently. It's much
easier to set this as the default, as it won't impact any non-screenshot
tests. Additionally, it also can't break the rendering of any existing
tests, because the rendered scene is identical, just the scene content
is translated. The added padding color also matches our tests in Parley.

For "display" screenshots, the padding having a background color isn't
great; we could consider making that optional or configurable.

We could also use a more "vibrant" padding color. E.g. Parley uses
`#a6c8ff`.