Put Vello's default features behind the default feature flag in transitive dependents (#1467)
This enables users of `xilem` and the `masonry` crates to opt out of
`vello`'s default features. As described in
[#1395](https://github.com/linebender/xilem/issues/1395) this is
particularly useful to opt out of `wgpu`'s default backends.

As `vello` is a transitive dependency to `xilem` and the `masonry`
crates it is necessary to change the default features in all of them. It
also enables use cases where a user would want to opt out of default
features of one of the crates in this repository if they were to
introduce default features in the future.