Add efficient child replacement methods to widgets. (#1494)
The only way to update children of `Flex`, `Grid`, `IndexedStack`, and
`ZStack` is with `push` / `insert` / `remove`. This is quite
inefficient. In order to replace the 0-index child, you have to first
call `remove` and then immediately combo it with `insert`. Both methods
are `O(n)` as they will shift the whole collection.

This PR introduces new `set_child` methods to these widgets, allowing
for `O(1)` replacement of children.