Update Cargo.lock (#1496)
I took a look at the lockfile recently and found that
[`megalodon`](https://github.com/h3poteto/megalodon-rs) > `oauth2 v4`
pulled in old versions of reqwest / hyper / http / h2. I submitted some
upgrade + cleanup PRs to `megalodon` and it's now got a new release with
`oauth v5` + current versions of reqwest and so on. Gets rid of a lot of
old stuff from `Cargo.lock` :)