Fix `TextInput` placeholder text alignment (#1504)
The placeholder text in `TextInput` was not respecting the
`text_alignment` property. So I added `text_alignment` field to
`TextInput`.

To show why this is a problem, this is my current version of Runebender
Xilem, note that the text input boxes are centered when they have
values, but I could not find a way to center the placeholder text for
the same input boxes.

<img width="1142" height="931" alt="Screenshot 2025-11-29 at 2 55 13 PM"
src="https://github.com/user-attachments/assets/91dc96a3-7e58-4cfc-8cdf-16174b3c828e"
/>