Add Canvas widget (#1519)
Alternative to #1506 (actually it builds on top of it again and applies
the review comments/suggestions, quite a group effort :))

I was bugged by the `Arc<dyn Fn>` and the way it was handled.

IMO it's a little bit better layered this way and more predictable when
the canvas draw function is actually changed, and allows the app state
in the canvas draw function by using `Fn(Arg<'_, State>, &mut Scene,
Size)` instead of just `Arc<dyn Fn(&mut Scene, Size)>` which would
require capturing in any realistic use-case (accessing the user state)
and thus the `Arc` would be somewhat useless, as always a new one needs
to be created.

Additionally I've added an example (as I think this is good practice
when adding new widgets).
(also to check whether this results in latency, i.e. when pressing the
button, that the scene is updated immediately instead of a frame-delay
(fortunately this doesn't result in frame-delay))

---------

Co-authored-by: Richard Dodd <richard.o.dodd@gmail.com>
Co-authored-by: Olivier FAURE <couteaubleu@gmail.com>
Co-authored-by: Myles Wirth <regime_skylark.0y@icloud.com>
Co-authored-by: Miles Wirht <114884788+philocalyst@users.noreply.github.com>
Co-authored-by: Daniel McNab <36049421+DJMcNab@users.noreply.github.com>
Co-authored-by: Muhammad Ragib Hasin <ragib.hasin@gmail.com>