Improve `Flex` child constraint accuracy. (#1540)
`Flex` contained a hack in the form of always having a phantom `0.0001`
flex factor present. This caused the child constraints to be awkward
numerically. Even if rounding could prevent issues in a lot of cases, it
still made debugging layout annoying due to getting e.g.
123.8567689512667 instead of 124.

This PR removes the phantom flex factor and guards against division by
zero explicitly.